backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0

# Phase 2
pypdf==4.2.0
trafilatura==1.9.0
qdrant-client==1.9.1
asyncpg==0.29.0

# Phase 3
langgraph==0.1.19
langchain-core==0.2.22

# Phase 4
mlflow==2.13.2

# Phase 5
redis==5.0.4
cohere==5.5.0